requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
Flask-SQLAlchemy==2.5.1
marshmallow==3.17.1
